requirements.txt
requests==2.26.0
flake8
PyJWT==2.1.0
isort==5.12.0
asgiref>=3.7.2
Django>=4.2.3
django-filter>=23.2
djangorestframework>=3.14.0
drf-base64>=2.0
fpdf>=1.7.2
gunicorn>=20.1.0
isort>=5.12.0
Pillow>=10.0.0
psycopg2-binary>=2.9.6
pytz>=2023.3
reportlab>=4.0.4
sqlparse>=0.4.4
python-dotenv>=1.0.0
djoser>=2.2.0